student_api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pyodbc==5.1.0
pandas==2.2.1
azure-identity==1.15.0